% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSMGPR.R
\name{kalman}
\alias{kalman}
\title{Kalman Filter Function}
\usage{
kalman(y, phi, Sigmav, Sigmaw, m0 = 0, Sigma0 = 1, n = 2)
}
\arguments{
\item{y}{Numeric vector of observations.}

\item{phi}{Numeric, the parameter phi in the Kalman filter.}

\item{Sigmav}{Numeric, the parameter Sigmav in the Kalman filter.}

\item{Sigmaw}{Numeric, the parameter Sigmaw in the Kalman filter.}

\item{m0}{Numeric, initial value of m.}

\item{Sigma0}{Numeric, initial value of Sigma.}

\item{n}{Numeric, number of future time points to predict.}
}
\value{
A list containing updated and predicted means and variances.
}
\description{
Implements a Kalman filter for time series data.
}
