% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSMGPR.R
\name{kf.gp}
\alias{kf.gp}
\title{Gaussian Process Regression with Kalman Filter}
\usage{
kf.gp(y, gamma, Sigmaw, m0 = 0, Sigma0 = 1, n = 2)
}
\arguments{
\item{y}{Numeric vector, the time series data to model.}

\item{gamma}{Numeric, the gamma parameter in GPR.}

\item{Sigmaw}{Numeric, the noise variance parameter.}

\item{m0}{Numeric, initial mean.}

\item{Sigma0}{Numeric, initial variance.}

\item{n}{Numeric, number of steps for prediction.}
}
\value{
A list with updated and predicted means and variances.
}
\description{
Applies Gaussian Process Regression (GPR) using a Kalman filter approach.
}
