% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSMGPR.R
\name{kf.loglikelihood1}
\alias{kf.loglikelihood1}
\title{Log Likelihood for Gaussian Process Regression}
\usage{
kf.loglikelihood1(y, mu.p, Sigma.p, Sigmaw, m0 = 0, Sigma0 = 1)
}
\arguments{
\item{y}{Numeric vector, the time series data.}

\item{mu.p}{Numeric vector, the predicted means.}

\item{Sigma.p}{Numeric vector, the predicted variances.}

\item{Sigmaw}{Numeric, the noise variance parameter.}

\item{m0}{Numeric, initial mean.}

\item{Sigma0}{Numeric, initial variance.}
}
\value{
Numeric, the log likelihood value.
}
\description{
Calculates the log likelihood for a given set of parameters in GPR.
}
