% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSMGPR.R
\name{optim_parm}
\alias{optim_parm}
\title{Optimize Parameters for Gaussian Process Regression}
\usage{
optim_parm(y)
}
\arguments{
\item{y}{Numeric vector, the time series data.}
}
\value{
A list containing optimal values for gamma and Sigmaw.
}
\description{
Finds optimal parameters for GPR using optimization techniques.
}
