% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pca}
\alias{pca}
\title{PCA Function}
\usage{
pca(data_matrix, cumulative_var = 0.9)
}
\arguments{
\item{scale}{PCA scaling}

\item{center}{PCA centering}
}
\value{
Result from PCA
}
\description{
PCA Function
}
