% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POpt.R
\name{port_opt}
\alias{port_opt}
\title{Optimal Portfolio Weights Calculation}
\usage{
port_opt(covar_mat, returns, num_components, lambda = 1)
}
\arguments{
\item{covar_mat}{A square, symmetric matrix representing the covariance matrix of asset returns.}

\item{returns}{A numeric vector of expected returns for each asset.}

\item{num_components}{An integer indicating the number of assets in the portfolio.}

\item{lambda}{A numeric value representing the risk aversion parameter (default is 1).}
}
\value{
A numeric vector of portfolio weights that minimizes the risk for a given level of return.
}
\description{
This function calculates the optimal portfolio weights based on a given covariance matrix, expected returns, and the number of components.
It uses a quadratic programming approach to find the solution.
}
\examples{
covar_mat <- matrix(c(0.1, 0.2, 0.2, 0.3), ncol=2)
returns <- c(0.1, 0.15)
num_components <- length(returns)
port_opt(covar_mat, returns, num_components)

}
